test: refactor: extract method to trait
